+++ src/configs/all.ts
@@ -4,6 +4,7 @@ export const all: Linter.BaseConfig = {
   plugins: ['filenames-simple'],
   rules: {
     'filenames-simple/extname': 'error',
+    'filenames-simple/named-export': 'error',
     'filenames-simple/naming-convention': 'error',
     'filenames-simple/no-index': 'error',
   },