+++ src/rules/index.ts
@@ -2,6 +2,7 @@ import { Rule } from 'eslint';
 
 import { casing } from './casing';
 import { extname } from './extname';
+import { namedExport } from './named-export';
 import { namingConvention } from './naming-convention';
 import { noIndex } from './no-index';
 
@@ -10,6 +11,7 @@ type Rules = { [key: string]: Rule.RuleModule | undefined };
 export const rules: Rules = {
   casing, // NOTE: Deprecated, I will remove this at v1.0.0
   extname,
+  'named-export': namedExport,
   'naming-convention': namingConvention,
   'no-index': noIndex,
 };