+++ CHANGELOG.md
@@ -1,10 +1,12 @@
 # Unreleased
 ## Features
 * [#9](https://github.com/epaew/eslint-plugin-filenames-simple/pull/9) Add new rule `no-index`
+* [#21](https://github.com/epaew/eslint-plugin-filenames-simple/pull/21) Add new rule `named-export`
 
 ## Others
 * [#10](https://github.com/epaew/eslint-plugin-filenames-simple/pull/10) Add CHANGELOG.md
 * [#17](https://github.com/epaew/eslint-plugin-filenames-simple/pull/17) Update package keywords
+* Generate type declaration files
 
 # 0.2.0
 ## Features